pyproject.toml
[project]
name = "crm-sales-bot"
version = "1.0.0"
description = "CRM Sales Bot with SPIN methodology - intelligent B2B sales assistant"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Core
    "faster-whisper>=1.1.0",
    "pypdf>=5.0.0",
    "requests>=2.32.0",
    "pyyaml>=6.0.2",
    "xlrd>=2.0.1",

    # NLP
    "pymorphy3>=2.0.6",

    # Numerical (embeddings served via TEI)
    "numpy>=2.1.0",

    # Async
    "aiohttp>=3.11.0",

    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # Validation & Logging
    "pydantic>=2.10.0",
    "structlog>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
]

llm = [
    # For running local LLM inference
    "vllm>=0.14.0",
    "transformers>=5.0.0",
    "torch>=2.6.0",
]

[project.scripts]
crm-bot = "src.bot:main"

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
